Get rid of JettyUtils in favor of Jetty specific JettyHttpUtils
