--- src/main/java/com/github/tomakehurst/wiremock/jetty/JettyUtils.java
@@ -1,45 +1,0 @@
-/*
- * Copyright (C) 2015-2024 Thomas Akehurst
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.github.tomakehurst.wiremock.jetty;
-
-public class JettyUtils {
-
-  private static final boolean IS_JETTY;
-
-  static {
-    // do the check only once per classloader / execution
-    IS_JETTY = isClassExist("org.eclipse.jetty.server.Request");
-  }
-
-  private JettyUtils() {
-    // Hide public constructor
-  }
-
-  public static boolean isJetty() {
-    return IS_JETTY;
-  }
-
-  private static boolean isClassExist(String type) {
-    try {
-      ClassLoader contextCL = Thread.currentThread().getContextClassLoader();
-      ClassLoader loader = contextCL == null ? JettyUtils.class.getClassLoader() : contextCL;
-      Class.forName(type, false, loader);
-      return true;
-    } catch (Exception e) {
-      return false;
-    }
-  }
-}
